feat: add new general/player configs
