--- src/config/togetherActivities.ts
@@ -1,50 +1,0 @@
-interface Activity {
-	label: string;
-	value: string;
-}
-export const activities: Activity[] = [
-	{
-		label: "Youtube",
-		value: "880218394199220334",
-	},
-	{
-		label: "Chess",
-		value: "832012774040141894",
-	},
-	{
-		label: "Poker",
-		value: "755827207812677713",
-	},
-	{
-		label: "Checkers",
-		value: "832013003968348200",
-	},
-	{
-		label: "Betrayal",
-		value: "773336526917861400",
-	},
-	{
-		label: "Fishington",
-		value: "814288819477020702",
-	},
-	{
-		label: "Words Snack",
-		value: "879863976006127627",
-	},
-	{
-		label: "Doodle Crew",
-		value: "878067389634314250",
-	},
-	{
-		label: "Spellcast",
-		value: "852509694341283871",
-	},
-	{
-		label: "Awkword",
-		value: "879863881349087252",
-	},
-	{
-		label: "Letter Tile",
-		value: "879863686565621790",
-	},
-];
